[feat]Refactered all routes in a separate file
