+++ TradePlay/src/router/requires-auth.jsx
@@ -1,12 +1,13 @@
-import { useLocation, Navigate } from "react-router-dom";
+import { useLocation, Navigate, Outlet } from "react-router-dom";
 import { useAuth } from "../context";
+import { Outlet } from "react-router-dom";
 
-const RequiresAuth = ({ children }) => {
+const RequiresAuth = () => {
   let location = useLocation();
   const { userDataState } = useAuth();
   const { token } = userDataState;
   return token ? (
-    children
+    <Outlet />
   ) : (
     <Navigate to="/login" state={{ from: location }} replace />
   );
