+++ TradePlay/src/utils/index.jsx
@@ -1,4 +1,2 @@
 export { changeTitle } from "./change-title";
-export { authHeader } from "./auth-header";
 export { inList } from "./in-list";
-export { saveApiDataToContext, changeCategory } from "./video-context-methods";
